+++ lib/common/common.dart
@@ -30,7 +30,6 @@ export 'package.dart';
 export 'path.dart';
 export 'permission.dart';
 export 'picker.dart';
-export 'pipe.dart';
 export 'preferences.dart';
 export 'print.dart';
 export 'protocol.dart';